Added name and updated addNumber and reversNumber
In the first branch,
a. Edit line 2 to have your name.
b. addNumber()
i. This function needs to be modified to add the argument variable
‘num’ to GitDemo’s ‘previousNumber’ variable.
c. reverseNumber()
i. Add a new “private double reverseNum;” variable to the GitDemo
class.
ii. Add the statement,
reverseNum = newNumber;
in the reverseNumber() function.
d. Commit the changes to the first branch. Be sure to write a commit